refactored Create report for Surgery Count (Type)  #22037 Signed-off-by: Pubudu <pubudupiyankara.me@gmail.com>
